feat: add email address
This might be a horrible idea, especially since I did not utilise any
preventitive measures. Here is why I am avoiding all the preventitive
measures I have read and can come up with. `-` indicates a measure, and
it is answered by the proceeding `>`, sort of like a `switch` statement
in C.

Measures and defenses:

- Use an image.

- Use Unicode characters that look similar to the corresponding Latin
characters.

- Just do not write the email address in full. For example, just say the
part before the "@" symbol.

> This means that `mailto` can not be used.

- Use JavaScript to write to the HTML document since apparently, most
bots do not run JavaScript.

> I would like to make JavaScript as optional as possible since some
people probably do not want to use JavaScript.

- Use some sort of forwarding feature, for example, with Cloudflare.

> I do have access to this via <bridgeduel.xyz>. However, it is
pointless if I want to reply to emails.

- Make an email address just for this.

> This email address is not being used for much so I think it is fine to
use it as essentially my public email address. In a way, I *am*
following this measure.

- Make a button to show the email.

> Yeah there is not really any defense against this. I just want to
experiment and see how bad it is for one to leave their email address in
plain sight.